+++ doc/whatsnew.rst
@@ -52,6 +52,9 @@ v0.2.8.dev
 
 - Add K-Nearest-Neighbor Regression: :class:`pyriemann.regression.KNearestNeighborRegressor`
 
+- Add example to compare classifiers on synthetic datasets
+
+
 v0.2.7 (June 2021)
 ------------------